@@ -11,21 +11,22 @@ description: Scanbot Barcode Scanner SDK Example App for Flutter
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.1.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: '>=2.0.1'
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Scanbot Barcode Scanner SDK
-  barcode_scanner: ^1.1.0
-
-  path_provider: ^1.6.27
-  image_picker: 0.6.3+1
-  cupertino_icons: ^0.1.2
+  barcode_scanner:
+    path: ../scanbot-barcode-sdk-flutter
+  path_provider: ^2.0.1
+  image_picker: ^0.7.3
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
