@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:barcode_scanner/barcode_scanning_data.dart';
 
 class BarcodeFormatItemWidget extends StatelessWidget {
-  BarcodeFormat format;
-  bool selected;
-  ValueChanged<bool> onSelect;
+  final BarcodeFormat format;
+  final bool selected;
+  final ValueChanged<bool?> onSelect;
 
   BarcodeFormatItemWidget(this.format, this.selected,
-      {@required this.onSelect});
+      {required this.onSelect});
 
   @override
   Widget build(BuildContext context) {
@@ -21,7 +21,7 @@ class BarcodeFormatItemWidget extends StatelessWidget {
             child: Row(
               children: <Widget>[
                 Text(
-                  barcodeFormatEnumMap[format],
+                  barcodeFormatEnumMap[format] ?? "",
                   style: TextStyle(inherit: true, color: Colors.black),
                 ),
                 Expanded(child: Container()),
